requirements/project.txt
boto3
s3fs
pandas
pyyaml
dbt-core
dbt-postgres
great-expectations